+++ src/basilisp/lang/compiler/generator.py
@@ -1,3 +1,6 @@
+# pylint: disable=too-many-lines
+
+import ast
 import collections
 import contextlib
 import functools
@@ -27,7 +30,6 @@ from typing import (
 
 import attr
 
-from basilisp import _pyast as ast
 from basilisp.lang import keyword as kw
 from basilisp.lang import list as llist
 from basilisp.lang import map as lmap
@@ -732,6 +734,7 @@ def expressionize(
     return ast.FunctionDef(
         name=fn_name,
         args=ast.arguments(
+            posonlyargs=[],
             args=args,
             kwarg=None,
             vararg=vargs,
@@ -925,6 +928,7 @@ def __deftype_classmethod_to_py_ast(
             node=ast.FunctionDef(
                 name=munge(node.name),
                 args=ast.arguments(
+                    posonlyargs=[],
                     args=list(
                         chain((ast.arg(arg=class_name, annotation=None),), fn_args)
                     ),
@@ -964,6 +968,7 @@ def __deftype_property_to_py_ast(
                 node=ast.FunctionDef(
                     name=method_name,
                     args=ast.arguments(
+                        posonlyargs=[],
                         args=list(
                             chain([ast.arg(arg=this_name, annotation=None)], fn_args)
                         ),
@@ -1123,6 +1128,7 @@ def __multi_arity_deftype_dispatch_method(
         node=ast.FunctionDef(
             name=name,
             args=ast.arguments(
+                posonlyargs=[],
                 args=[ast.arg(arg=method_prefix, annotation=None)],
                 kwarg=None,
                 vararg=ast.arg(arg=_MULTI_ARITY_ARG_NAME, annotation=None),
@@ -1205,6 +1211,7 @@ def __deftype_method_arity_to_py_ast(
                 node=ast.FunctionDef(
                     name=method_name if method_name is not None else munge(arity.name),
                     args=ast.arguments(
+                        posonlyargs=[],
                         args=list(
                             chain((ast.arg(arg=this_name, annotation=None),), fn_args)
                         ),
@@ -1253,6 +1260,7 @@ def __deftype_staticmethod_to_py_ast(
             node=ast.FunctionDef(
                 name=munge(node.name),
                 args=ast.arguments(
+                    posonlyargs=[],
                     args=fn_args,
                     kwarg=None,
                     vararg=varg,
@@ -1645,6 +1653,7 @@ def __single_arity_fn_to_py_ast(
                         py_fn_node(
                             name=py_fn_name,
                             args=ast.arguments(
+                                posonlyargs=[],
                                 args=fn_args,
                                 kwarg=None,
                                 vararg=varg,
@@ -1821,6 +1830,7 @@ def __multi_arity_dispatch_fn(  # pylint: disable=too-many-arguments,too-many-lo
                 py_fn_node(
                     name=name,
                     args=ast.arguments(
+                        posonlyargs=[],
                         args=[],
                         kwarg=None,
                         vararg=ast.arg(arg=_MULTI_ARITY_ARG_NAME, annotation=None),
@@ -1895,6 +1905,7 @@ def __multi_arity_fn_to_py_ast(  # pylint: disable=too-many-locals
                 py_fn_node(
                     name=arity_name,
                     args=ast.arguments(
+                        posonlyargs=[],
                         args=fn_args,
                         kwarg=None,
                         vararg=varg,
@@ -2442,6 +2453,7 @@ def _reify_to_py_ast(
             ast.FunctionDef(
                 name="meta",
                 args=ast.arguments(
+                    posonlyargs=[],
                     args=[
                         ast.arg(arg="self", annotation=None),
                     ],
@@ -2458,6 +2470,7 @@ def _reify_to_py_ast(
             ast.FunctionDef(
                 name="with_meta",
                 args=ast.arguments(
+                    posonlyargs=[],
                     args=[
                         ast.arg(arg="self", annotation=None),
                         ast.arg(arg="new_meta", annotation=None),
@@ -3275,6 +3288,7 @@ def _const_meta_kwargs_ast(
 
 
 @_const_val_to_py_ast.register(bool)
+@_const_val_to_py_ast.register(bytes)
 @_const_val_to_py_ast.register(type(None))
 @_const_val_to_py_ast.register(complex)
 @_const_val_to_py_ast.register(float)
@@ -3467,6 +3481,7 @@ def _const_record_to_py_ast(form: IRecord, ctx: GeneratorContext) -> GeneratedPy
     form_seq = runtime.to_seq(form)
     assert form_seq is not None, "IRecord types must be iterable"
 
+    # pylint: disable=no-member
     keys, vals = [], []
     vals_deps: List[ast.AST] = []
     for k, v in form_seq:
@@ -3527,7 +3542,7 @@ def _const_type_to_py_ast(form: IType, ctx: GeneratorContext) -> GeneratedPyAST:
 
     ctor_args = []
     ctor_arg_deps: List[ast.AST] = []
-    for field in attr.fields(tp):  # type: ignore[arg-type]
+    for field in attr.fields(tp):  # type: ignore[arg-type, misc, unused-ignore]
         field_nodes = _const_val_to_py_ast(getattr(form, field.name, None), ctx)
         ctor_args.append(field_nodes.node)
         ctor_args.extend(field_nodes.dependencies)